post-service/requirements.txt
grpcio
cassandra-driver
protobuf>=6.30.1
python-dotenv
pytest
pytest-mock
pytest-cov
